+++ src/Pet4U.Application/UseCases/UpdateSocialNetworks/IUpdateSocialNetworksHandler.cs
@@ -1,7 +1,7 @@
 using Pet4U.Application.UseCases.CreateVolunteer;
 using Pet4U.Domain.Shared;
 
-namespace Pet4U.Application.UseCases.UpdateMainInfo;
+namespace Pet4U.Application.UseCases.UpdateSocialNetworks;
 
 public interface IUpdateSocialNetworks
 {
